requirements.txt
streamlit
opencv-python
easyocr
matplotlib
numpy
pillow
